+++ packages/eslint-plugin-wpcalypso/tests/lib/rules/import-docblock.js
@@ -1,0 +1,44 @@
+/**
+ * @fileoverview Enforce external, internal dependencies docblocks
+ * @author Automattic
+ * @copyright 2016 Automattic. All rights reserved.
+ * See LICENSE.md file in root directory for full license.
+ */
+'use strict';
+
+//------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+const rule = require( '../../../lib/rules/import-docblock' ),
+	RuleTester = require( 'eslint' ).RuleTester;
+
+//------------------------------------------------------------------------------
+// Tests
+//------------------------------------------------------------------------------
+
+( new RuleTester( {
+	parserOptions: {
+		ecmaVersion: 6,
+		sourceType: 'module'
+	}
+} ) ).run( 'import-docblock', rule, {
+	valid: [
+		{
+			code:
+`/**
+ * External dependencies
+ */
+import eslint from \'eslint\';`
+		}
+	],
+
+	invalid: [
+		{
+			code: 'import eslint from \'eslint\';',
+			errors: [ {
+				message: 'Missing external, internal dependencies docblocks'
+			} ]
+		}
+	]
+} );